# [Frontend] Speedup frontend test