--- a/requirements/test.txt
+++ b/requirements/test.txt
@@ -177,9 +177,10 @@ hiredis==3.0.0
     # via tensorizer
 httpcore==1.0.6
     # via httpx
-httpx==0.27.2
+httpx==0.28.1
     # via
     #   -r requirements/test.in
+    #   pytest-httpx
     #   schemathesis
 huggingface-hub==0.30.1
     # via
@@ -502,6 +503,7 @@ pytest==8.3.3
     #   genai-perf
     #   pytest-asyncio
     #   pytest-forked
+    #   pytest-httpx
     #   pytest-mock
     #   pytest-rerunfailures
     #   pytest-shard
@@ -511,6 +513,8 @@ pytest==8.3.3
 pytest-asyncio==0.24.0
     # via -r requirements/test.in
 pytest-forked==1.6.0
+    # via -r requirements/test.in
+pytest-httpx==0.35.0
     # via -r requirements/test.in
 pytest-mock==3.14.0
     # via genai-perf
@@ -651,9 +655,7 @@ six==1.16.0
     #   rfc3339-validator
     #   rouge-score
 sniffio==1.3.1
-    # via
-    #   anyio
-    #   httpx
+    # via anyio
 sortedcontainers==2.4.0
     # via hypothesis
 soundfile==0.12.1