… tests/v1/entrypoints/openai/test_completion_non_streaming.py
… tests/v1/entrypoints/openai/test_streaming.py