FILL IN THE PR DESCRIPTION HERE

FIX https://github.com/vllm-project/vllm/issues/15369 cc @aarnphm 

This PR speeds up the OpenAI completion API tests by mocking the HTTP layer and initializing the engine directly, without spinning up a full server. For non-streaming requests, we use pytest-httpx to mock HTTP responses. However, since pytest-httpx does not support streaming responses, streaming requests are handled by directly invoking the engine logic, bypassing the HTTP layer.

This approach reduces the runtime of test_completion.py on my machine from 58s to 24s.

If this method proves robust, we can extend it to additional frontend tests to further improve test performance.

Note:
- This is an initial attempt to assess whether this approach is maintainable and reliable for broader adoption.
- Feedback and suggestions for better or more idiomatic ways to mock streaming and non-streaming OpenAI endpoints are welcome.

Before: 58s
![Screenshot from 2025-05-17 23-47-44](https://github.com/user-attachments/assets/509268fe-c346-4115-94c1-1d0d995887cc)
After: 24s
![Screenshot from 2025-05-18 00-58-22](https://github.com/user-attachments/assets/c9eafebc-bcb5-487b-b594-76605fa22352)

<!--- pyml disable-next-line no-emphasis-as-heading -->
**BEFORE SUBMITTING, PLEASE READ <https://docs.vllm.ai/en/latest/contributing/overview.html>** (anything written below this line will be removed by GitHub Actions)